+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 2026-03-21
+
+## Added
+
+- vscode terminal tab closes when tmux window exits
+- exiting vscode terminal tab kills tmux window
+- tmux-integrated.autoConnect option. Connect to windows when opening workspace (default: true)
+- Update node version from 18 to 20 to enable auto publish workflow
+
 ## [0.1.5] - 2026-03-20
 
 ### Fixed